Always load comments on a conversation page
If an element with the id of "disqus_thread" exists on the page we load
the disqus embed directly. We also support having the old style of a
fetch more div with a link to the comments (which will load comments on
click).

We might want to diffrentiate this on mobil so we do not automaticly
load comments.